[authority] Move reconfiguration state into PerEpochStore
The reconfiguration state information is tightly coupled to other epoch tables. For example, we do state transaction from CloseUserCerts => CloseAllCerts based on `epoch_tables.end_of_publish`.

If external event happens, for example due to state sync where we need to change an epoch any state regarding reconfiguration of previous epoch becomes irrelevant.

I did not think super well yet what to do with the RwLock holding the ReconfigurationState in regards with bigger picture. I think most likely the `epoch_tables` itself needs to be stored in the RwLock. (will do that in separate PR when I have more clarity on that)

Note that `open_all_certs` method is also deleted in this case - there is no way to transition back to open reconfiguration state, other than by switching to a new epoch(that will create empty reconfiguration state table).
