+++ crates/sui-core/src/authority/authority_store.rs
@@ -40,8 +40,6 @@ pub struct CertLockGuard(LockGuard);
 const NUM_SHARDS: usize = 4096;
 const SHARD_SIZE: usize = 128;
 
-const RECONFIG_STATE_INDEX: u64 = 0;
-
 /// ALL_OBJ_VER determines whether we want to store all past
 /// versions of every object in the store. Authority doesn't store
 /// them, but other entities such as replicas will.
@@ -989,29 +987,6 @@ impl<S: Eq + Debug + Serialize + for<'de> Deserialize<'de>> SuiDataStore<S> {
         };
 
         Ok(assigned_seq)
-    }
-
-    pub fn load_reconfig_state(&self) -> SuiResult<ReconfigState> {
-        let state = match self
-            .perpetual_tables
-            .reconfig_state
-            .get(&RECONFIG_STATE_INDEX)?
-        {
-            Some(state) => state,
-            None => {
-                let init_state = ReconfigState::default();
-                self.store_reconfig_state(&init_state)?;
-                init_state
-            }
-        };
-        Ok(state)
-    }
-
-    pub fn store_reconfig_state(&self, new_state: &ReconfigState) -> SuiResult {
-        self.perpetual_tables
-            .reconfig_state
-            .insert(&RECONFIG_STATE_INDEX, new_state)?;
-        Ok(())
     }
 
     /// This function is called at the end of epoch for each transaction that's
