+++ crates/sui-core/src/consensus_adapter.rs
@@ -188,6 +188,8 @@ impl ConsensusAdapter {
         #[allow(clippy::collapsible_if)] // This if can be collapsed but it will be ugly
         if self
             .authority
+            .database
+            .epoch_tables()
             .get_reconfig_state_read_lock_guard()
             .is_reject_user_certs()
         {
@@ -287,8 +289,9 @@ impl ConsensusAdapter {
     /// It transition reconfig state to reject new certificates from user
     /// ConsensusAdapter will send EndOfPublish message once pending certificate queue is drained
     pub fn close_epoch(self: &Arc<Self>) -> SuiResult {
+        let epoch_tables = self.authority.database.epoch_tables();
         let send_end_of_publish = {
-            let reconfig_guard = self.authority.get_reconfig_state_write_lock_guard();
+            let reconfig_guard = epoch_tables.get_reconfig_state_write_lock_guard();
             let pending_certificates = self.pending_certificates.lock();
             let send_end_of_publish = pending_certificates.is_empty();
             self.authority.close_user_certs(reconfig_guard);
@@ -313,8 +316,9 @@ impl ConsensusAdapter {
         self: &Arc<Self>,
         transaction: ConsensusTransaction,
     ) -> SuiResult<JoinHandle<()>> {
+        let epoch_tables = self.authority.database.epoch_tables();
         let _lock = if transaction.is_user_certificate() {
-            let lock = self.authority.get_reconfig_state_read_lock_guard();
+            let lock = epoch_tables.get_reconfig_state_read_lock_guard();
             if !lock.should_accept_user_certs() {
                 return Err(SuiError::ValidatorHaltedAtEpochEnd);
             }
@@ -391,7 +395,8 @@ impl ConsensusAdapter {
         }
         let send_end_of_publish =
             if let ConsensusTransactionKind::UserTransaction(certificate) = &transaction.kind {
-                let reconfig_guard = self.authority.get_reconfig_state_read_lock_guard();
+                let epoch_tables = self.authority.database.epoch_tables();
+                let reconfig_guard = epoch_tables.get_reconfig_state_read_lock_guard();
                 // note that pending_certificates lock is always acquired *after* reconfiguration lock
                 // acquiring locks in different order might lead to deadlocks
                 let mut pending_certificates = self.pending_certificates.lock();
